+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPT9PFXT4KAQ7XWKEX4SRSND"
+        "01KPTE09PYM92GB8BEMYW6RV4B"
     )
     r"""The unique identifier of the data source"""