+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPT9PFYJAK0P6JTKAFWDJ70G"
+        "01KPTE09Q32XCG0HZC9EJCM5E6"
     )
     r"""The unique identifier of the data source"""